Adding separated classes for authorization and default Adding redirect function to Authorization class
