+++ app/Controllers/CategoryController.php
@@ -2,17 +2,16 @@
 
 namespace App\Controllers;
 
-use App\Core\Controller;
+use App\Core\Authorization;
 use App\Models\Category;
 use App\Models\Offer;
 use App\Models\Subcategory;
 
-class CategoryController extends Controller
+class CategoryController extends Authorization
 {
     public function index(): void
     {
-        header("Location: ".DIRPAGE);
-        exit;
+        $this->redirect(DIRPAGE);
     }
 
     public function show(string $slug = ""): void
@@ -32,8 +31,7 @@ class CategoryController extends Controller
         $this->setKeywords("ofertas, produtos, preço, {$categoryInfo['name']}");
 
         if (empty($slug)) {
-            header("Location: ".DIRPAGE);
-            exit;
+            $this->redirect(DIRPAGE);
         }
 
         $subcategoryId = null;
@@ -52,8 +50,7 @@ class CategoryController extends Controller
                 $categoryId,
                 "category")
             ) {
-                header("Location: ".DIRPAGE."category/show/{$slug}");
-                exit;
+                $this->redirect(DIRPAGE."category/show/{$slug}");
             }
         }
