+++ app/Models/Ability.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Core\Table;
+
 class Ability extends Table
 {
     public function __construct()
