+++ app/Models/Offer.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Core\Table;
+
 class Offer extends Table
 {
     public function __construct()
