+++ app/Models/Role.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Core\Table;
+
 class Role extends Table
 {
     public function __construct()
