+++ app/Models/User.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Core\Table;
+
 class User extends Table
 {
     public function __construct()
